% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnorm.R
\name{fnorm}
\alias{fnorm}
\title{The Frobenius Norm}
\usage{
fnorm(tnsr)
}
\arguments{
\item{tnsr, }{a 3-mode tensor S3 class object}
}
\value{
The Frobenius norm
}
\description{
The Frobenius norm of an array is the square root of the sum of its squared elements. This function works for vector and matrix arguments as well.
}
\examples{
T <- t_rand(modes=c(2,2,4))
fnorm(T$data)
}
\references{
Friedland, S., & Aliabadi, M. (2018). Linear algebra and matrices. Society for Industrial and Applied Mathematics.
}
\author{
Kyle Caudle

Randy Hoover

Jackson Cates

Everett Sandbo
}
