\name{raytrace}
\alias{raytrace}
\docType{data}
\title{
 Subset of raytrace data
}
\description{
4 tensors (128 x 128 x 128) for 4 different gray scale images.  boat, flashlight, keyboard, scooter.
}
\usage{data("raytrace")}
\format{
  The format is:

    raytrace$boat

    raytrace$flashlight

    raytrace$keyboard

    raytrace$scooter

}

\references{
Hoover RC, Braman KS, Hao N (2011b). “Pose estimation from a single image using tensor
decomposition and an algebra of circulants.” In 2011 IEEE/RSJ International Conference
on Intelligent Robots and Systems, pp. 2928–2934. IEEE.
}
\examples{
data(raytrace)
}
\keyword{datasets}
