% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_rand.R
\name{t_rand}
\alias{t_rand}
\title{Create a random tensor}
\usage{
t_rand(modes = c(3, 4, 5))
}
\arguments{
\item{modes, }{the 3 modes of the output Tensor}
}
\value{
an S3 Tensor object
}
\description{
Generate a Tensor with specified modes whose entries are iid normal(0,1).
}
\examples{
 t_rand(c(4,4,4))
}
\references{
Imported from rTensor2 package version 2.0.0.
}
\author{
Kyle Caudle

Randy Hoover

Jackson Cates

Everett Sandbo
}
