% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constr.R
\name{Tensor}
\alias{Tensor}
\title{Creates an S3 class for a tensor}
\usage{
Tensor(data, x, y, z)
}
\arguments{
\item{data}{Numeric numbers in the tensor}

\item{x, }{mode 1 dimension}

\item{y, }{mode 2 dimension}

\item{z, }{mode 3 dimension}
}
\value{
S3 class tensor
}
\description{
Creates an S3 class for a tensor
}
