% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tINVdct.R
\name{tINVdct}
\alias{tINVdct}
\title{Performs inverse of 3-mode tensor using the discrete cosine transform.}
\usage{
tINVdct(tnsr)
}
\arguments{
\item{tnsr, }{a 3-mode S3 tensor class object}
}
\value{
S3 class tensor
#' @examples
T <- t_rand(modes=c(2,2,4))
print(tINVdct(T))
}
\description{
Performs inverse of 3-mode tensor using the discrete cosine transform.
}
\author{
Kyle Caudle

Randy Hoover

Jackson Cates

Everett Sandbo
}
