% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tINVdht.R
\name{tINVdht}
\alias{tINVdht}
\title{Performs inverse of 3-mode tensor using the discrete Hadley transform.}
\usage{
tINVdht(tnsr)
}
\arguments{
\item{tnsr, }{a 3-mode S3 tensor class object}
}
\value{
S3 class tensor
}
\description{
Performs inverse of 3-mode tensor using the discrete Hadley transform.
}
\examples{
T <- t_rand(modes=c(2,2,4))
print(tINVdht(T))
}
\author{
Kyle Caudle

Randy Hoover

Jackson Cates

Everett Sandbo
}
