% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tINVdst.R
\name{tINVdst}
\alias{tINVdst}
\title{Performs inverse of 3-mode tensor using the discrete sine transform.}
\usage{
tINVdst(tnsr)
}
\arguments{
\item{tnsr, }{a 3-mode S3 tensor class object}
}
\value{
S3 class tensor
}
\description{
Performs inverse of 3-mode tensor using the discrete sine transform.
}
\examples{
T <- t_rand(modes=c(2,2,4))
print(tINVdst(T))
}
\author{
Kyle Caudle

Randy Hoover

Jackson Cates

Everett Sandbo
}
