% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tINVdwht.R
\name{tINVdwht}
\alias{tINVdwht}
\title{Performs inverse of 3-mode tensor using the discrete Walsh Hadley transform.}
\usage{
tINVdwht(tnsr)
}
\arguments{
\item{tnsr, }{a 3-mode S3 tensor class object}
}
\value{
S3 class tensor
}
\description{
Performs inverse of 3-mode tensor using the discrete Walsh Hadley transform.
}
\examples{
T <- t_rand(modes=c(2,2,4))
print(tINVdwht(T))
}
\author{
Kyle Caudle

Randy Hoover

Jackson Cates

Everett Sandbo
}
