% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tINVdwt.R
\name{tINVdwt}
\alias{tINVdwt}
\title{Performs inverse of 3-mode tensor using the discrete wavelet transform.}
\usage{
tINVdwt(tnsr)
}
\arguments{
\item{tnsr, }{a 3-mode S3 tensor class object}
}
\value{
S3 class tensor
}
\description{
Performs inverse of 3-mode tensor using the discrete wavelet transform.
}
\examples{
tnsr <- t_rand(modes=c(2,2,4))
print(tINVdwt(tnsr))
}
\author{
Kyle Caudle

Randy Hoover

Jackson Cates

Everett Sandbo
}
