% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tINVfft.R
\name{tINVfft}
\alias{tINVfft}
\title{Performs inverse of 3-mode tensor using the discrete fast fourier transform.}
\usage{
tINVfft(tnsr)
}
\arguments{
\item{tnsr, }{a 3-mode S3 tensor class object}
}
\value{
S3 class tensor
}
\description{
Performs inverse of 3-mode tensor using the discrete fast fourier transform.
}
\examples{
T <- t_rand(modes=c(2,2,4))
print(tINVfft(T))
}
\author{
Kyle Caudle

Randy Hoover

Jackson Cates

Everett Sandbo
}
