% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmean.R
\name{tmean}
\alias{tmean}
\title{Determines the mean of a 3D tensor along mode 2}
\usage{
tmean(tnsr)
}
\arguments{
\item{tnsr, }{a 3D tensor of dimensions n1,n2,n3}
}
\value{
S3 tensor class object
}
\description{
Determines the mean of a 3D tensor along mode 2
}
\examples{
tnsr <- t_rand(modes=c(3,4,5))
tmean(tnsr)
}
\references{
M. E. Kilmer, C. D. Martin, and L. Perrone, “A third-order generalization of the matrix svd as a product of third-order tensors,” Tufts University, Department of Computer Science, Tech. Rep. TR-2008-4, 2008

K. Braman, "Third-order tensors as linear operators on a space of matrices", Linear Algebra and its Applications, vol. 433, no. 7, pp. 1241-1253, 2010.
}
\author{
Kyle Caudle

Randy Hoover

Jackson Cates

Everett Sandbo
}
