% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tSVDdwht.R
\name{tSVDdwht}
\alias{tSVDdwht}
\title{Singular value decomposition (SVD) of a 3D tensor using the discrete Walsh Hadley transform}
\usage{
tSVDdwht(tnsr)
}
\arguments{
\item{tnsr, }{a 3-mode S3 tensor class object}
}
\value{
U, the left singular value tensor object (\eqn{m} x \eqn{m} x \eqn{k})

V, The right singular value tensor object (\eqn{n} x \eqn{n} x \eqn{k})

S: A diagonal tensor (\eqn{m} x \eqn{n} x \eqn{k})#' @examples
 V: The right singular value tensor object (\eqn{n} x \eqn{n} x \eqn{k})
S: A diagonal tensor (\eqn{m} x \eqn{n} x \eqn{k})
}
\description{
Singular value decomposition (SVD) of a 3D tensor using the discrete Walsh Hadley transform
}
\examples{
T <- t_rand(modes=c(2,2,4))
tSVDdwht(T)
}
\references{
M. E. Kilmer, C. D. Martin, and L. Perrone, “A third-order generalization of the matrix svd as a product of third-order tensors,” Tufts University, Department of Computer Science, Tech. Rep. TR-2008-4, 2008

K. Braman, "Third-order tensors as linear operators on a space of matrices", Linear Algebra and its Applications, vol. 433, no. 7, pp. 1241-1253, 2010.
}
\author{
Kyle Caudle

Randy Hoover

Jackson Cates

Everett Sandbo
}
